requirements.txt
Django==2.2.2
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2==2.8.2
pytz==2019.1
sorl-thumbnail==12.5.0
sqlparse==0.3.0
